# BGDIINF_SB-1541 index model fields