--- a/app/stac_api/views.py
+++ b/app/stac_api/views.py
@@ -44,6 +44,13 @@ def get_collection_etag(request, *args, **kwargs):
     The ETag is an UUID4 computed on each object changes (including relations; provider and links)
     '''
     tag = get_etag(Collection.objects.filter(name=kwargs['collection_name']))
+    # uncomment the following lines to see the SQL query and the EXPLAIN.. ANALYZE
+    # output:
+    # logger.debug(
+    #     Collection.objects.filter(name=kwargs['collection_name']
+    #                              ).explain(verbose=True, analyze=True)
+    # )
+    # logger.debug(Collection.objects.filter(name=kwargs['collection_name']).query)
     return tag
 
 
@@ -55,6 +62,16 @@ def get_item_etag(request, *args, **kwargs):
     tag = get_etag(
         Item.objects.filter(collection__name=kwargs['collection_name'], name=kwargs['item_name'])
     )
+    # uncomment the following lines to see the SQL query and the EXPLAIN.. ANALYZE
+    # output:
+    # logger.debug(
+    #     Item.objects.filter(collection__name=kwargs['collection_name'],
+    #                         name=kwargs['item_name']).explain(verbose=True, analyze=True)
+    # )
+    # logger.debug(
+    #     Item.objects.filter(collection__name=kwargs['collection_name'],
+    #                         name=kwargs['item_name']).query
+    # )
     return tag
 
 
@@ -64,6 +81,15 @@ def get_asset_etag(request, *args, **kwargs):
     The ETag is an UUID4 computed on each object changes
     '''
     tag = get_etag(Asset.objects.filter(item__name=kwargs['item_name'], name=kwargs['asset_name']))
+    # uncomment the following lines to see the SQL query and the EXPLAIN.. ANALYZE
+    # output:
+    # logger.debug(
+    #     Asset.objects.filter(item__name=kwargs['item_name'],
+    #                          name=kwargs['asset_name']).explain(verbose=True, analyze=True)
+    # )
+    # logger.debug(
+    #     Asset.objects.filter(item__name=kwargs['item_name'], name=kwargs['asset_name']).query
+    # )
     return tag
 
 
@@ -176,7 +202,10 @@ class ItemsList(generics.GenericAPIView, views_mixins.CreateModelMixin):
 
         if date_time:
             queryset = queryset.filter_by_datetime(date_time)
-
+        # uncomment the following lines to see the SQL query and the EXPLAIN.. ANALYZE
+        # output:
+        # logger.debug(queryset.explain(verbose=True, analyze=True))
+        # logger.debug(queryset.query)
         return queryset
 
     def list(self, request, *args, **kwargs):
@@ -234,6 +263,10 @@ class ItemDetail(
         # filter based on the url
         queryset = Item.objects.filter(collection__name=self.kwargs['collection_name']
                                       ).prefetch_related('assets', 'links')
+        # uncomment the following lines to see the SQL query and the EXPLAIN.. ANALYZE
+        # output:
+        # logger.debug(queryset.explain(verbose=True, analyze=True))
+        # logger.debug(queryset.query)
         return queryset
 
     def get_write_request_data(self, request, *args, partial=False, **kwargs):
@@ -293,7 +326,10 @@ class SearchList(generics.GenericAPIView, mixins.ListModelMixin):
                 queryset = queryset.filter_by_query(dict_query)
             if 'intersects' in query_param:
                 queryset = queryset.filter_by_intersects(json.dumps(query_param['intersects']))
-
+        # uncomment the following lines to see the SQL query and the EXPLAIN.. ANALYZE
+        # output:
+        # logger.debug(queryset.explain(verbose=True, analyze=True))
+        # logger.debug(queryset.query)
         return queryset
 
     def list(self, request, *args, **kwargs):