… app/stac_api/migrations/0002_auto_20210216_1527.py
… app/stac_api/migrations/0002_auto_20210218_0726.py