* added indices where extensive tests identified them to be useful
* added debug prints for SQL query and output of `EXPLAIN.. ANALYZE`