@hansmannj 2021-01-29
In order to check the influence of model field indices I made 10 queries like:
```bash
curl --request POST \
  --url http://0.0.0.0:8000/api/stac/v0.9/search \
  --header 'Content-Type: application/json' \
  --data '{"query": 
 	{"datetime":
	 	{"lte":"2525-01-02T00:00:00Z"}  
	},
 "limit": 100
}'
```

one without indices, one with indices.
There was not a significant difference, and if any, then the queries w/o the indices were even slightly faster:

`SQLPanel_sql_time`, average of 10 queries:
w/o indices:  118.3ms (201 queries), with indices: 124.6 ms (204 queries)

`TimerPanel_stime` (System CPU time):
w/o indices: 265.5 ms, with indices: 288.2 ms