@hansmannj 2021-01-29
However, this query uses the ` item_dttme_start_end_dttm_idx` index:
```bash
curl --request POST \
  --url http://0.0.0.0:8000/api/stac/v0.9/search \
  --header 'Content-Type: application/json' \
  --data '{"query":
		{"datetime":
				{"in":["1960-10-28T13:05:10Z","2525-10-19T00:00:00Z"]}
		}
}'
```

But for some reason, no items are returned (found). Unfortunately I do not have time to debug further today and most likely not during the next week in my holidays. So could continue the week after, or someone can continue here, in case it is urgent.