@ltclm 2021-02-12
Salut @hansmannj 

the btree index seems to be the right choice for datetime columns and the operators that we are using for the filter.
I did some tests with 250'000 items:

```sql
CREATE INDEX idx_stac_api_item_created
    ON public.stac_api_item USING btree
    (created ASC NULLS LAST)
    TABLESPACE pg_default; 

select  count(*), min(created), max(created) FROM stac_api_item;
250000	"2020-02-13 14:22:56.246497+00"	"2021-02-12 14:22:16.891535+00"

-- with btree index on created column
EXPLAIN ANALYZE select * FROM stac_api_item WHERE created > '2020-01-01' -- seq scan --> alle daten		-- Execution Time: 61.298 ms
EXPLAIN ANALYZE select * FROM stac_api_item WHERE created > '2020-07-01' -- seq scan --> alle daten		-- Execution Time: 56.135 ms
EXPLAIN ANALYZE select * FROM stac_api_item WHERE created > '2020-12-01' -- index --> nicht alle daten		-- Execution Time: 34.665 ms
EXPLAIN ANALYZE select * FROM stac_api_item WHERE created > '2021-07-01' -- index --> keine daten		-- Execution Time: 0.019 ms
EXPLAIN ANALYZE select * FROM stac_api_item WHERE created > '2021-07-01' -- index --> keine daten		-- Execution Time: 0.021 ms

-- without btree index on created column
EXPLAIN ANALYZE select * FROM stac_api_item WHERE created > '2020-01-01' -- seq scan --> alle daten		-- Execution Time: 63.226 ms
EXPLAIN ANALYZE select * FROM stac_api_item WHERE created > '2020-07-01' -- seq scan --> alle daten		-- Execution Time: 57.670 ms
EXPLAIN ANALYZE select * FROM stac_api_item WHERE created > '2020-12-01' -- seq scan  --> nicht alle daten	-- Execution Time: 45.828 ms
EXPLAIN ANALYZE select * FROM stac_api_item WHERE created > '2021-07-01' -- seq scan  --> keine daten		-- Execution Time: 119.938 ms
EXPLAIN ANALYZE select * FROM stac_api_item WHERE created > '2021-07-01' -- seq scan  --> keine daten		-- Execution Time: 112.852 ms


```

Whether the index is used or not depends on the datetime filter you are using, if the filter is covering the whole dataset. a sequential scan will be done in any case and the index is not used. see first two examples. In any other cases an index scan will be done.

This is the impact of this index on the write performance (250'000 items with sql):

```sql
-- without the index
INSERT 0 250000
Query returned successfully in 11 secs 261 msec.

-- with index
INSERT 0 250000
Query returned successfully in 13 secs 550 msec.
```