@hansmannj 2021-02-15
> Salut @hansmannj
> 
> the btree index seems to be the right choice for datetime columns and the operators that we are using for the filter.
> I did some tests with 250'000 items:
> 

Cool, thanks for testing @ltclm!
In the special case of `created` I also thought about using BRIN, as the data for that column should be sorted and in theory also never change, once created. But after reading this: https://info.crunchydata.com/blog/avoiding-the-pitfalls-of-brin-indexes-in-postgres I am not so sure about using BRIN any more..